Fix all 5 AI voice pipeline bugs from deep review
Bug #1 (CRITICAL): AI thread consumed input BEFORE acquiring converter
lock. When try_lock failed, audio was silently dropped. Fix: use
blocking lock() and only pop_slice after lock is held.

Bug #2 (CRITICAL): Context trimming used new buffer length instead of
old context length, producing wrong trim ratios. Fix: save old context
length before overwriting.

Bug #3 (IMPORTANT): VoiceConverter was created with hardcoded 48kHz but
engine uses device's actual sample rate. Fix: read actual_sample_rate()
from engine and pass to VoiceConverterConfig.

Bug #4 (IMPORTANT): Generator output (40kHz native) was length-matched
via linear interpolation instead of proper resampling. Fix: resample
from GENERATOR_SAMPLE_RATE (40kHz) to device rate via rubato.

Bug #5: parseInt truncated fractional semitones on pitch slider.
Fix: use parseFloat.

+++ ui/app.js
@@ -68,7 +68,7 @@ async function init() {
   // AI controls
   document.getElementById('unload-btn').addEventListener('click', unloadVoice);
   document.getElementById('ai-pitch').addEventListener('input', (e) => {
-    const val = parseInt(e.target.value);
+    const val = parseFloat(e.target.value);
     document.getElementById('ai-pitch-value').textContent = `${val >= 0 ? '+' : ''}${val} st`;
     invoke('set_ai_pitch', { semitones: val }).catch(console.error);
   });
